fix: implement missing planned optimizations lost during git reset
- removeTxFromSubtrees O(1) lookup: use SubtreeIndex from DiskTxMap for
  direct subtree access instead of scanning all 1000 subtrees (falls back
  to linear scan when DiskTxMap is not active)
- processCompleteSubtree: update SubtreeIndex in Badger for all txs when
  subtree is chained, enabling the O(1) lookup above
- resetSubtreeState: use currentTxMap.Clear() instead of replacing with
  NewSplitTxInpointsMap, preserving DiskTxMap when active
- lastValidatedBlocks cache: add WithEvictionFunction callback to close
  mmap-backed subtrees when blocks expire from the 2-minute cache
- SubtreeProcessingBatch.Close(): new method to release mmap subtree
  resources; called after batch completion in both sequential and pipeline
  processing paths in quick_validate.go
- clearRecentMapsForDisk: fix uint16 vs int type inconsistency

+++ services/blockassembly/subtreeprocessor/disk_tx_map.go
@@ -386,7 +386,7 @@ func (m *DiskTxMap) flushAllDisks() {
 // clearRecentMapsForDisk clears recent maps for filter shards mapped to a specific disk.
 func (m *DiskTxMap) clearRecentMapsForDisk(diskIdx int) {
 	for i := range m.shards {
-		if int(uint16(i)%uint16(m.numDisks)) != diskIdx {
+		if i%m.numDisks != diskIdx {
 			continue
 		}
 		m.shards[i].mu.Lock()
